Turn arkwork example into tests
