+++ examples/rust/virgo/src/lib.rs
@@ -1,0 +1,1 @@
+mod api;
