+++ wrappers/rust/icicle-core/src/virgo/mod.rs
@@ -24,6 +24,29 @@ impl<'a> VirgoConfig<'a> {
     }
 }
 
+#[repr(C)]
+#[derive(Debug, Clone)]
+pub struct MerkleTreeConfig<'a> {
+    /// Details related to the device such as its id and stream id. See [DeviceContext](@ref device_context::DeviceContext).
+    pub ctx: DeviceContext<'a>,
+    pub MAX_MIMC_K: u32,
+}
+
+impl<'a> Default for MerkleTreeConfig<'a> {
+    fn default() -> Self {
+        Self::default_for_device(DEFAULT_DEVICE_ID)
+    }
+}
+
+impl<'a> MerkleTreeConfig<'a> {
+    pub fn default_for_device(device_id: usize) -> Self {
+        Self {
+            ctx: DeviceContext::default_for_device(device_id),
+            MAX_MIMC_K: 128,
+        }
+    }
+}
+
 /////////////
 
 pub trait Virgo<F: FieldImpl> {
@@ -52,6 +75,14 @@ pub trait Virgo<F: FieldImpl> {
 
     fn bk_produce_case_2(
         config: &VirgoConfig,
+        table: &HostOrDeviceSlice<F>,
+        result: &mut HostOrDeviceSlice<F>,
+        n: u32,
+    ) -> IcicleResult<()>;
+
+    //// Merkle tree
+    fn build_merkle_tree(
+        config: &MerkleTreeConfig,
         table: &HostOrDeviceSlice<F>,
         result: &mut HostOrDeviceSlice<F>,
         n: u32,
@@ -112,6 +143,19 @@ where
     <<F as FieldImpl>::Config as Virgo<F>>::bk_produce_case_2(config, table, result, n)
 }
 
+pub fn build_merkle_tree<F>(
+    config: &MerkleTreeConfig,
+    table: &HostOrDeviceSlice<F>,
+    result: &mut HostOrDeviceSlice<F>,
+    n: u32,
+) -> IcicleResult<()>
+where
+    F: FieldImpl,
+    <F as FieldImpl>::Config: Virgo<F>,
+{
+    <<F as FieldImpl>::Config as Virgo<F>>::build_merkle_tree(config, table, result, n)
+}
+
 #[macro_export]
 macro_rules! impl_virgo {
     (
@@ -121,7 +165,7 @@ macro_rules! impl_virgo {
         $field_config:ident
       ) => {
         mod $field_prefix_ident {
-            use crate::virgo::{$field, $field_config, CudaError, DeviceContext, VirgoConfig};
+            use crate::virgo::{$field, $field_config, CudaError, DeviceContext, MerkleTreeConfig, VirgoConfig};
 
             extern "C" {
                 #[link_name = concat!($field_prefix, "BkSumAllCase1")]
@@ -159,6 +203,16 @@ macro_rules! impl_virgo {
                 #[link_name = concat!($field_prefix, "BkProduceCase2")]
                 pub(crate) fn _bk_produce_case_2(
                     config: &VirgoConfig,
+                    arr: *const $field,
+                    result: *mut $field,
+                    n: u32,
+                ) -> CudaError;
+            }
+
+            extern "C" {
+                #[link_name = concat!($field_prefix, "BuildMerkleTree")]
+                pub(crate) fn _build_merkle_tree(
+                    config: &MerkleTreeConfig,
                     arr: *const $field,
                     result: *mut $field,
                     n: u32,
@@ -224,6 +278,17 @@ macro_rules! impl_virgo {
             ) -> IcicleResult<()> {
                 unsafe {
                     $field_prefix_ident::_bk_produce_case_2(config, table.as_ptr(), result.as_mut_ptr(), n).wrap()
+                }
+            }
+
+            fn build_merkle_tree(
+                config: &MerkleTreeConfig,
+                table: &HostOrDeviceSlice<$field>,
+                result: &mut HostOrDeviceSlice<$field>,
+                n: u32,
+            ) -> IcicleResult<()> {
+                unsafe {
+                    $field_prefix_ident::_build_merkle_tree(config, table.as_ptr(), result.as_mut_ptr(), n).wrap()
                 }
             }
         }
