+++ wrappers/rust/icicle-curves/icicle-bn254/src/virgo/mod.rs
@@ -3,6 +3,7 @@ use crate::curve::{ScalarCfg, ScalarField};
 use icicle_core::error::IcicleResult;
 use icicle_core::impl_virgo;
 use icicle_core::traits::IcicleResultWrap;
+use icicle_core::virgo::MerkleTreeConfig;
 use icicle_core::virgo::Virgo;
 use icicle_core::virgo::VirgoConfig;
 use icicle_cuda_runtime::device_context::DeviceContext;
